requirements.txt
azure-cognitiveservices-speech
azure-identity
openai